fixed download button
