+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@aws-sdk/client-s3": "^3.485.0",
+    "@aws-sdk/s3-request-presigner": "^3.485.0",
     "aws-sdk": "^2.1532.0",
     "next": "14.0.4",
     "next-themes": "^0.2.1",
