+++ pages/api/downloadCv.tsx
@@ -1,0 +1,34 @@
+import type { NextApiRequest, NextApiResponse } from "next";
+import { S3Client, GetObjectCommand } from "@aws-sdk/client-s3";
+import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
+
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  const s3 = new S3Client({
+    region: process.env.AWS_REGION,
+    credentials: {
+      accessKeyId: process.env.AWS_ACCESS_KEY_ID ?? "",
+      secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY ?? "",
+    },
+  });
+
+  const zipKey = "Curriculum-Vitae.zip";
+
+  try {
+    const command = new GetObjectCommand({
+      Bucket: process.env.S3_BUCKET_NAME,
+      Key: zipKey,
+    });
+
+    const url = await getSignedUrl(s3, command, {
+      expiresIn: 60,
+    });
+
+    res.status(200).json({ url });
+  } catch (error) {
+    console.error("Error generating pre-signed URL:", error);
+    res.status(500).json({ error: "Error generating URL" });
+  }
+}
